added the fixes for dhruva's garbage

+++ model/profiles.py
@@ -8,9 +8,9 @@ class Profile(db.Model):
     __tablename__ = 'profiles'
     _username = db.Column(db.String(80), unique=True, nullable=False, primary_key=True)
     _email = db.Column(db.String(120), unique=True, nullable=False)
-    _high_high_score = db.Column(db.Integer(500), unique=False, nullable=False)
-    _starred_games = db.Column(db.Integer(100), unique=True, nullable=False)
-    _points_per_second = db.Column(db.Integer(100), unique=False, nullable=False)
+    _high_high_score = db.Column(db.Integer, unique=False, nullable=False)
+    _starred_games = db.Column(db.Integer, unique=True, nullable=False)
+    _points_per_second = db.Column(db.Integer, unique=False, nullable=False)
 
     def __init__(self, username, email, high_high_score, starred_games, points_per_second):
         self._username = username
